Fixed bug using Host header, and working on Java tests.
The Host header was not being used correctly to set the links sent back in
hypermedia.  I can get each test in StubServerFunctionalTest to pass on its
own, but not when run together, for the same reason that the ControlServerFunctionalTest
fails -- the stub.close() call doesn't appear to work in time.  I thought
I fixed that by not responding until the Node server close event fired, but
that apparently hasn't een enough